+++ src/locus_http_download.erl
@@ -92,13 +92,11 @@
 -export_type([event_request_sent/0]).
 
 -type event_download_dismissed() ::
-    {download_dismissed, {http, response_status(), headers(), body()}}.
+    {download_dismissed, full_http_response()}.
 -export_type([event_download_dismissed/0]).
 
 -type event_download_failed_to_start() ::
-    {download_failed_to_start, {http, response_status(), headers(), body()}} |
-    {download_failed_to_start, {error, term()}} |
-    {download_failed_to_start, timeout}.
+    {download_failed_to_start, reason_for_download_failing_to_start()}.
 -export_type([event_download_failed_to_start/0]).
 
 -type event_download_started() ::
@@ -110,6 +108,16 @@
     {download_finished, BodySize :: non_neg_integer(), {error, term()}} |
     {download_finished, BodySize :: non_neg_integer(), {error, timeout}}.
 -export_type([event_download_finished/0]).
+
+-type reason_for_download_failing_to_start() ::
+    full_http_response() |
+    {error, term()} |
+    timeout.
+-export_type([reason_for_download_failing_to_start/0]).
+
+-type full_http_response() ::
+    {http, response_status(), headers(), body()}.
+-export_type([full_http_response/0]).
 
 -type success() ::
     #{ headers := headers(),